feat: Refactor OSS Insight components and improve data handling
